Fixes for Mac build

+++ cocos/2d/CCFontFreeType.cpp
@@ -55,7 +55,7 @@ typedef struct _DataRef
 
 static std::unordered_map<std::string, DataRef> s_cacheFontData;
 
-FontFreeType * FontFreeType::create(const std::string &fontName, int fontSize, GlyphCollection glyphs, const char *customGlyphs,bool distanceFieldEnabled /* = false */,int outline /* = 0 */)
+FontFreeType * FontFreeType::create(const std::string &fontName, float fontSize, GlyphCollection glyphs, const char *customGlyphs,bool distanceFieldEnabled /* = false */,int outline /* = 0 */)
 {
     FontFreeType *tempFont =  new FontFreeType(distanceFieldEnabled,outline);
 
@@ -123,7 +123,7 @@ FontFreeType::FontFreeType(bool distanceFieldEnabled /* = false */,int outline /
     }
 }
 
-bool FontFreeType::createFontObject(const std::string &fontName, int fontSize)
+bool FontFreeType::createFontObject(const std::string &fontName, float fontSize)
 {
     FT_Face face;
     // save font name locally
